Menu: Avoid scrolling the entire page when the menu opens
When the menu has not been opened yet, floating-ui sometimes positions
it in a weird place off-screen. When the menu opens, we attempt to
scroll the highlighted item into view before floating-ui repositions the
menu in a reasonable place, and this can cause the page to scroll
wildly, typically to the very bottom.

Work around this by further deferring the scroll using
requestAnimationFrame(). Using nextTick() doesn't work here for some
reason, and in fact the scroll is already behind two nextTick()s, so
it's odd that it happens before the floating-ui repositioning at all.

Bug: T382505
Change-Id: I0e992c3854c70f8563275ebc58d471cf6f588900