[DEPRECATING CHANGE] Throw deprecation warnings for old APIs
In https://phabricator.wikimedia.org/T368443, props for the following
components were changed to add a boolean prop to turn on a feature
instead of relying on a string prop related to that feature:
- Dialog: `useCloseButton` prop added instead of relying on
  `closeButtonLabel` prop to enable the close button
- Field, Label: `optional` prop added instead of relying on
  `optionalFlag` prop to show the optional flag
- Message: `allowUserDismiss` prop added instead of relying on
  `dismissButtonLabel` prop to show the dismiss button
- SearchInput, TypeaheadSearch: `useButton` prop added instead of
  relying on the `buttonLabel` prop to show the search button

With these changes, we set up support for the old API where the
string props were used to both enable the feature and set up a
string label. However, we never informed Codex users that they
need to update their code nor added deprecation warnings. This
patch adds deprecationg warnings for the old APIs so that we
can eventually remove them in the next major version of Codex.

Bug: T390776
Change-Id: Ic8b0845f0b06d30a43e01aa4c083fa3d1753ab81