styles, sandbox: apply typography mixin to Sandbox demo pages
Apply the typography mixin to all Sandbox pages via DemoBaseLayout
component. The mixin is called at the top-level of the Less styles
and compiles the CSS to a more manageable specificity such as `h1`.
If the mixin call was made within the `<body>` element, then the
selector's specificity would increase and compiled to `body h1`.

Change-Id: Iffb2c115ab4ba3ca1b8a793e1045726628774c62