docs: Improve Lookup and MultiselectLookup demos with fetching
In particular, the Lookup demos have caused some issues
downstream as people have copied them and experienced
unexpected results. The LookupWithFetch demo was originally
created before we added the optional `inputValue` prop and
had an old way of internally tracking the input value that
is no longer needed.

This patch ensures that:
- All of these demos use the `inputValue` prop and its associated
  `update:input-value` event to model the ideal setup.
- The various refs are explained to help avoid further confusion.
- For Lookup, when an item is selected, new menu items are fetched
  to match the new input value.

Additionally, this patch fixes a bug in the MultiselectLookupField
demo where the input was case-sensitive.

Change-Id: I248694ca6b474ce15c455b05876ade819a726339