build: Update build-if-missing script in icons package
This script is meant to build the package only if it needs to be
built. The previous test just tested whether the `dist/` directory
existed, which meant the library would not be build if, say, a
new icon was added but the developer had built the package
locally in the past. This often caused issues when developers
would try to serve the docs site locally, causing errors when
an icon was missing.

This updates the script to compare the number of images in
the `src/` and `dist/` directories (once the existence of the
latter has been established). If the numbers differ, the package
will be built.

Change-Id: I4a0e57ed19ff59a21c525cdb1084f2a814dc58e7