docs: add favicon
This patch adds a 20x20 SVG type favicon to the doc site via
the head Vitepress configuration. The favicon is a Wikimedia
blue version of the Codex logo icon to ensure the icon is
visible in light and dark browser tabs.

Known issue: The favicon SVG does not display in Safari. A
favicon.ico file will work for most browsers.

Bug: T398375
Change-Id: I9a9b81a52b3f2b6bff5629d6ec55e1d5ac723760