TypeaheadSearch, styles: Improve footer styles for mobile view
The mobile view with thumbnails shares some styles with the
desktop version but not all. Notably, the footer styles should be
the same, so that the footer icon is aligned with the center of
the menu item thumbnails, but the input should not have additional
padding.

This patch adds a new dynamic class, `cdx-typeahead-search--is-mobile-view`,
and splits the `--show-thumbnail` styles between desktop-only and
desktop + mobile.

This also fixes a regression where the footer text wasn't aligned
with the menu item text, by adding back in padding-right on the
icon.

Bug: T399489
Change-Id: I488647d2215241622f59e3827a3df7db60786313