docs: Fix number props and type links with brackets
A bug in the code that processes default values for props was
causing the default value of the Breadcrumb's `maxVisible` prop
to display as "[]" rather than the actual number. This patch
fixes that vulnerability.

This patch also simplifies the code that handles array types,
which have brackets. It ensures that types that link to the
Types and Constants page only include the type name in the
link, followed by un-linked brackets. See the Menu component's
Vue usage docs for examples of different types of properties
handled by this code.

Finally, this patch adds a check for arrays of different types
(e.g. `(MenuItemData|MenuGroupData)[]`) to ensure that the
string contains opening and closing parentheses followed by
brackets.

Change-Id: Ic66890d5f28ddc97518b4facf52cdca2c220a8d0