build: Update 'svglint' and 'svgo' dependencies
 svglint                      3.1.0  →   4.1.0
 svgo                         3.3.2  →   4.0.0

Minimally changing the SVGO config to be compatible with v4.0.0+.
Gladly, the authors decided to remove the
`removeTitle` and `removeViewBox` plugins as default enabled options,
so we can remove the overrides.
Also `convertColors``, now converts all references to colors by default.
Lastly, running the optimization is changing the color in newly added
favicon.svg to lowercase as expected.

Bug: T399678
Change-Id: I46ba36b01ae4c13b01a7c78a262a98fd5c531115