useFloatingMenu: Pass in reference HTMLElement to autoUpdate() directly
If the reference element was a template ref, we were passing the ref to
autoUpdate(). However, this broke certain floating-ui features, causing
the menu not to stay anchored to the reference element when its
container was scrolled. Unwrapping the ref and passing in the
HTMLElement itself fixes this.

Also add an early return so that we can simplify existence checks.

Bug: T397786
Change-Id: I978fd1a193ca1d8f789e3b5e787348520702c4cf