build: Upgrade eslint-config-wikimedia from 0.28.2 to 0.31.0
This has been written to be the minimal change to actual code,
instead disabling almost all the newly-failing rules. Most of
said rules have come from the merger of Wikimedia and WMDE TS
rules, and are in an early stage, so feedback about how much
they're appropriate is very welcome!

Change-Id: I8323ea24b3880355232065118e5d196480af8c7f