docs: Apply HTML "`<tagged-element>` element" writing convention
Improving developer experience by unifying on a way to
refer to HTML elements. And mark them explicitly as code for
supporting assistive technology to read them out as code.
This is a mostly Claude induced alternative to a semi-manual,
semi-complex search and replace done in I4d77337.

Also fixing a number of typos, spellings and similar small issues.

Bug: T382574
Change-Id: I01c652d3059f128372b392fc35de9b7d5f55b372