docs: Add support for Matomo pageview analytics
We're going to use Matomo (formerly known as Piwik) to analyze which
pages are viewed how often on the Wikimedia-hosted version of the docs
site at doc.wikimedia.org.

This change supports that by injecting a tracking script from the URL in
the CODEX_MATOMO_URL environment variable. If the environment variable
isn't set, nothing changes and no pageview analytics are gathered.

Bug: T368970
Change-Id: I02f701c05497ee10de949dbee5d4255555dfa15b