TypeaheadSearch: Fix unavoidable deprecation warning in mobile view mode
When the TypeaheadSearch component was used with the isMobileView prop
set, we would throw a deprecation warning from SearchInput, but there
was nothing the caller of TypeaheadSearch could do to avoid it.

Fix this by not passing the buttonLabel prop to SearchInput if we're
also going to set the useButton prop to false.

Change-Id: I3f48d4308346347ff29fce16756e8c42aa1143ab