docs: Set up teleport target for docs site components
We had per-demo teleport targets for demo components, but we didn't have
a teleport target for components on the docs site outside of demos. This
didn't matter at first, because only Dialogs and Popovers were
teleported and they weren't used outside of demos, but now Menus are
also teleported and they are used by the docs site directly (in
IconLookup for example).

Create a teleport target for these menus and set `dir="ltr"` on it so
that direction-specific styles don't break. Keep the per-demo teleport
targets, but simplify the code setting those up; we no longer need to
wait to provide them until mount time because CdxTeleportTarget now
accepts template refs too.

Change-Id: Ia073159ee74029b5296494f80b0068ba6c6264f0