styles, Grid: Use `gap` shorthand for both, rows and columns
Sensible default for all Codex Grid usages, also in accordance
with original design specs.
It's more often than not, that the grid items don't come with
pre-set `margin`s, so a default row gap is needed as well.
Making the mixin as flexible as possible to define f.e. 0 row gaps
for adding the Grid mixin to an already established interface with
`margin`s set in place.
Note, the current flexible approach comes with caveat of not providing
fine-tuning for tablet and mobile gutters, but that seems like a rare
edge case at the time being and against design specs, while the missing
0 row gaps could actually slow down implementation.

Bug: T405372
Change-Id: If469949032d66c3e08652f8b5d2754cfc70a905f