Dialog: Make Dialog full screen on small screens
Also change how the footer is put at the bottom so that the body doesn't
have to grow to the full height of the Dialog in full screen mode. This
avoids a bug where the ResizeObserver never fires because the body
height never changes, causing the dividers between the header, body and
footer to never appear.

Change-Id: Id362b980c68c90389886f636d2b84c4ddb7ca790
Bug: T405015
Co-authored-by: Roan Kattouw <rkattouw@wikimedia.org>