Accordion: Unbreak markup in titles and descriptions
If you set the title or description of an Accordion to something like
`Foo <em>bar</em> baz`, it would display as `Foobarbaz`. This happened
because we set `display: flex;` on the title and description spans, but
there was no reason for that. (The title and description are arranged
using flexbox, but that's done by setting `display: flex` on their
parent.) Change this to `display: block`, which fixes the display of
marked-up text but doesn't alter the layout otherwise.

Change-Id: I7dc2fcdff48c18949f00addb19d451ed431f087c