Dialog: Always show close button in full-screen mode
In order to do this, always render the close button (and by extension,
the header), and hide it using CSS.

Also remove the extra padding on the body when the header isn't shown,
since the header is now always shown. Happily, the top+bottom padding on
the header is exactly equal to the amount of padding we were adding to
the body, so there's no visual difference.

Change-Id: Ia50128a97fa8d3dc6b9a7a16f12c52479d764b4b